+++ src/main/java/com/aetherteam/aether/item/accessories/cape/CapeItem.java
@@ -36,10 +36,10 @@ public class CapeItem extends AccessoryItem implements AccessorySlotProvider {
      */
     @Override
     public SlotTypeReference getSlotType() {
-        return getStaticIdentifier();
+        return getStaticSlotType();
     }
 
-    public static SlotTypeReference getStaticIdentifier() {
+    public static SlotTypeReference getStaticSlotType() {
         return AetherConfig.COMMON.use_default_accessories_menu.get() ? () -> "cape" : AetherAccessorySlots.getCapeSlotType();
     }
 }