+++ src/main/java/com/aetherteam/aether/item/accessories/pendant/PendantItem.java
@@ -45,10 +45,10 @@ public class PendantItem extends AccessoryItem implements AccessorySlotProvider 
      */
     @Override
     public SlotTypeReference getSlotType() {
-        return getStaticIdentifier();
+        return getStaticSlotType();
     }
 
-    public static SlotTypeReference getStaticIdentifier() {
+    public static SlotTypeReference getStaticSlotType() {
         return AetherConfig.COMMON.use_default_accessories_menu.get() ? () -> "necklace" : AetherAccessorySlots.getPendantSlotType();
     }
 }